feat(adsp-service-spring-sdk): allow sub-class and customization of security configuration
