+++ libs/adsp-service-spring-sdk/src/main/java/ca/ab/gov/alberta/adsp/sdk/access/AccessIssuer.java
@@ -2,7 +2,7 @@ package ca.ab.gov.alberta.adsp.sdk.access;
 
 import ca.ab.gov.alberta.adsp.sdk.tenant.Tenant;
 
-final class AccessIssuer {
+public final class AccessIssuer {
   private final Tenant tenant;
   private final boolean isCore;
